\name{dinos}
\alias{dinos}
\docType{data}
\title{Dinosaur data from Campione and Evans 2012}
\description{Humeral and Femoral circumference data for eight quadrupedal dinosaurs}
\usage{data(dinos)}
\format{
  A data frame with 8 observations and the following 3 variables.
  \describe{
    \item{\code{Taxon}}{a factor with levels \code{Brachiosaurs} \code{Corythosaurus} \code{Diplodocus} \code{Iguanodon} \code{Protoceratops} \code{Stegosaurus} \code{Styracosaurus} \code{Triceratops}}
    \item{\code{HC}}{a numeric vector of humeral circumferences}
    \item{\code{FC}}{a numeric vector of femoral circumferences}
  }
}
\source{
Campione, N. E. and Evans, D. C. (2012) A universal scaling relationship between body mass and proximal limb bone dimensions in quadrupedal terrestrial tetrapods. BMC Biology, 10, 60.
}