\name{MASSTIMATE-package}
\alias{MASSTIMATE}
\docType{package}
\title{Body mass estimation equations for vertebrates}
\description{Estimation equations are from a variety of sources but are, in general, based on regressions between skeletal measurements (e.g., femoral circumference) and body mass in living taxa}
\details{
\tabular{ll}{
Package: \tab MASSTIMATE\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2013-11-05\cr
License: \tab GPL(>= 2)\cr
}
}
\author{
Nicolas E. Campione

Maintainer: Nicolas E. Campione <nicolas.campione@geo.uu.se>
}
\references{
Anderson, J. F., Hall-Martin, A. and Russell, D.A. (1985) Long-bone circumference and weight in mammals, birds and dinosaurs. Journal of the Zoological Society of London A, 207, 53-61.

Campbell Jr., K.E. and Marcus, L. (1992) The relationships of hindlimb bone dimensions to body weight in birds. Natural History Museum of Los Angeles County Science Series, 36, 395-412.

Campione, N. E. and Evans, D. C. (2012) A universal scaling relationship between body mass and proximal limb bone dimensions in quadrupedal terrestrial tetrapods. BMC Biology, 10, 60.

Campione, N. E., Evans, D. C., Brown, C. M. and Carrano, M. T. (in prep) A mathematically-derived equation for estimating body mass in terrestrial bipedal tetrapods.

Christiansen, P. and Fari??a, R. A. (2004) Mass prediction in theropod dinosaurs. Historical Biology, 16, 85-92.

Mazzetta, G. V., Christiansen, P. and Fari??a, R. A. (2004) Giants and bizzares: body size of some southern South American Cretaceous dinosaurs. Historical Biology, 16, 71-83.
}
\keyword{ package }