\name{dinosbip}
\alias{dinosbip}
\docType{data}
\title{Dinosaur data from Campione et al. (in prep)}
\description{Femoral circumference data for 34 bipedal dinosaurs}
\usage{data(dinosbip)}
\format{
  A data frame with 34 observations and the following 3 variables.
  \describe{
    \item{\code{Taxon}}{a factor with the taxon names}
    \item{\code{FC}}{a numeric vector of femoral circumferences}
    \item{\code{cor}}{a numeric vector with the correction factor values to be used in 'cQE'}
  }
}
\source{
  Campione, N. E., Evans, D. C., Brown, C. M. and Carrano, M. T. (in prep) A mathematically-derived equation for estimating body mass in terrestrial bipedal tetrapods.
}