\name{MASSTIMATE-package}
\alias{MASSTIMATE}
\docType{package}
\title{Body mass estimation equations for vertebrates}
\description{Estimation equations are from a variety of sources and associated error estimation.}
\details{
\tabular{ll}{
Package: \tab MASSTIMATE\cr
Type: \tab Package\cr
Version: \tab 2.0\cr
Date: \tab 2020-09-01\cr
License: \tab GPL(>= 2)\cr
}
}
\author{
Nicolas E. Campione

Maintainer: Nicolas E. Campione <ncampion@une.edu.au>
}
\references{
Anderson, J. F., Hall-Martin, A. and Russell, D.A. (1985) Long-bone circumference and weight in mammals, birds and dinosaurs. Journal of the Zoological Society of London A, 207, 53-61.

Campbell Jr., K.E. and Marcus, L. (1992) The relationships of hindlimb bone dimensions to body weight in birds. Natural History Museum of Los Angeles County Science Series, 36, 395-412.

Campione, N. E. and Evans, D. C. (2012) A universal scaling relationship between body mass and proximal limb bone dimensions in quadrupedal terrestrial tetrapods. BMC Biology, 10, 60.

Campione, N. E., Evans, D. C., Brown, C. M. and Carrano, M. T. (2014) Body mass estimation in non-avian bipeds using a theoretical conversion to quadruped stylopodial proportions. Methods in Ecology and Evolution, 5(9), 913-923.

Campione, N. E. (2017) Extrapolation of a universal scaling relationship for estimating body masses in extinct terrestrial vertebrates. Paleobiology, 43, 693-699.

Campione, N. E. and Evans, D. C. (2020) The accuracy and precision of body mass estimation in non-avian dinosaurs. Biological Reviews.

Christiansen, P. and Farina, R. A. (2004) Mass prediction in theropod dinosaurs. Historical Biology, 16, 85-92.

Erickson G. M. and Tumanova, T. A. (2000). Growth curve of Psittacosaurus mongoliensis Orborn (Ceratopsia; Psittacosauridae) inferred from long bone histology. Zoological Journal of the Linnean Society, 130, 551-566.

Mazzetta, G. V., Christiansen, P. and Farina, R. A. (2004) Giants and bizzares: body size of some southern South American Cretaceous dinosaurs. Historical Biology, 16, 71-83.
}
\keyword{package}